Seperated models into individual files for more readability
